fix(ai-accounts): stop presenting unrefreshed quota as current
The AI accounts panel could show quota values that were days old while
looking completely live. Three separate defects stacked up:

1. A failed probe kept quota_json untouched but still advanced
   upstream_checked_at and version, so the stored payload and the
   timestamp describing it drifted apart. Four accounts in production are
   currently in this state; the oldest stopped being confirmed 16 days ago
   and its values are still served as-is.

2. Upstreams report only a subset of windows on many probes (codex answers
   with just rate_limit or just additional_rate_limits in ~20% of them).
   Replacing quota_json wholesale therefore dropped the other window until
   a later probe happened to include it again.

3. A 200 that parsed to zero windows silently overwrote the payload with
   an empty set, and nothing in the probe path logged either failure mode.

Quota freshness is now modelled explicitly. Each window carries its own
observed_at and windows are merged by quota_key: confirmed windows are
refreshed and restamped, absent ones are carried forward with their
original timestamp, and windows unconfirmed past 24h are dropped. Row-level
quota_observed_at dates the payload as a whole and, unlike
upstream_checked_at, only moves when something was actually observed. The
migration backfills it from ai_account_subject_quota_points, which is only
written after a successful probe and therefore dates existing rows
correctly even for long-failing accounts.

Probe failures and empty payloads now log with subject/provider/auth_index.

Files over the size ratchet were split rather than grown: subject status
storage, the status view assembly, and the runtime schema constant each
moved to their own file, leaving all three baselines smaller.

Verified: ./scripts/ci-pr.sh (gofmt, secret scan, structure check, go vet,
go test ./..., golangci-lint, go build).
