+++ internal/management/aiaccountstatus/service.go
@@ -12,6 +12,7 @@ import (
 	managementapitools "github.com/router-for-me/CLIProxyAPI/v6/internal/management/apitools"
 	"github.com/router-for-me/CLIProxyAPI/v6/internal/usage"
 	coreauth "github.com/router-for-me/CLIProxyAPI/v6/sdk/cliproxy/auth"
+	log "github.com/sirupsen/logrus"
 	"golang.org/x/sync/singleflight"
 )
 
@@ -477,6 +478,13 @@ func (s *Service) refreshOne(jobID, tenantID string, auth *coreauth.Auth, subjec
 		probe.SubscriptionSource = asserted.SubscriptionSource
 	}
 	if probeErr != nil {
+		// Probe outcomes were previously invisible in logs, so "the panel shows a
+		// week-old number" could not be traced to the upstream failure behind it.
+		log.WithFields(log.Fields{
+			"auth_subject_id": subjectID,
+			"provider":        auth.Provider,
+			"auth_index":      auth.Index,
+		}).WithError(probeErr).Warn("ai account status probe failed; stored quota is kept but not refreshed")
 		_ = usage.UpdateAIAccountSubjectProbeFailure(subjectID, auth.Provider, "probe_failed", probeErr.Error(), checked)
 		_ = usage.UpdateAIAccountRefreshFailure(
 			tenantID, subjectID, auth.Index, auth.Provider, string(auth.Status),
@@ -504,6 +512,16 @@ func (s *Service) refreshOne(jobID, tenantID string, auth *coreauth.Auth, subjec
 			r.UpdatedAt = checked
 		})
 		return
+	}
+	// A 200 that parses to zero windows means the upstream payload shape changed or
+	// the account genuinely has no quota. Either way the stored windows are carried
+	// forward unrefreshed, so surface it instead of silently serving aging values.
+	if len(probe.Quotas) == 0 {
+		log.WithFields(log.Fields{
+			"auth_subject_id": subjectID,
+			"provider":        auth.Provider,
+			"auth_index":      auth.Index,
+		}).Warn("ai account status probe returned no quota windows; stored quota is kept but not refreshed")
 	}
 	s.applyRuntimeQuotaProbe(ctx, auth, probe)
 
@@ -864,90 +882,6 @@ func (s *Service) ListStatus(tenantID string, authIndexes, authSubjectIDs []stri
 		items = append(items, statusViewFromSharedRecord(row, auth, summaries[sid], subject, bindingCounts[sid]))
 	}
 	return StatusListResponse{Items: items}, nil
-}
-
-func statusViewFromSharedRecord(row usage.AIAccountSubjectStatusRecord, auth *coreauth.Auth, summary usage.AuthSubjectUsageSummary, subject usage.AIAccountSubjectRecord, bindingCount int) AccountStatusView {
-	view := AccountStatusView{
-		AuthSubjectID: row.AuthSubjectID, Provider: row.Provider, StatusScope: usage.AIAccountStatusScopeShared,
-		SubjectScope: subject.SubjectScope, ShareEligible: subject.ShareEligible, SubjectSeedKind: subject.SeedKind,
-		CurrentTenantBindingCount: bindingCount, RefreshState: row.LastProbeState, HealthStatus: row.HealthStatus,
-		PlanType: row.PlanType, RestrictionSummary: row.RestrictionSummary, ErrorSummary: row.ErrorSummary,
-		ErrorCode: row.ErrorCode, Quotas: row.Quotas, ResetCreditCount: row.ResetCreditCount,
-		ResetCreditExpirations: row.ResetCreditExpirations, Usage: summary,
-		SubscriptionStartedAt: row.SubscriptionStartedAt, SubscriptionExpiresAt: row.SubscriptionExpiresAt,
-		SubscriptionSource: row.SubscriptionSource, UpstreamCheckedAt: row.UpstreamCheckedAt,
-		ExpiresAt: row.SubscriptionExpiresAt, Version: row.Version, UpdatedAt: timePointer(row.UpdatedAt),
-	}
-	if auth != nil {
-		view.AuthIndex = auth.Index
-		if view.Provider == "" {
-			view.Provider = auth.Provider
-		}
-		if view.HealthStatus == "" {
-			view.HealthStatus = string(auth.Status)
-		}
-	}
-	if view.Quotas == nil {
-		view.Quotas = []usage.QuotaWindowDTO{}
-	}
-	if view.Usage.AuthSubjectID == "" {
-		view.Usage.AuthSubjectID = row.AuthSubjectID
-	}
-	if !summary.UpdatedAt.IsZero() {
-		view.UsageUpdatedAt = timePointer(summary.UpdatedAt)
-	}
-	if view.Usage.WeeklyQuotaUsed == nil && auth != nil {
-		view.Usage.WeeklyQuotaUsed = weeklyUsedFromQuotas(row.Quotas, primaryWeeklyKeys(auth.Provider)...)
-	}
-	return view
-}
-
-func timePointer(value time.Time) *time.Time {
-	if value.IsZero() {
-		return nil
-	}
-	value = value.UTC()
-	return &value
-}
-
-func weeklyUsedFromQuotas(quotas []usage.QuotaWindowDTO, preferred ...string) *float64 {
-	pref := make(map[string]struct{}, len(preferred))
-	for _, k := range preferred {
-		pref[k] = struct{}{}
-	}
-	for i := range quotas {
-		q := &quotas[i]
-		if q.Percent == nil {
-			continue
-		}
-		if len(pref) > 0 {
-			if _, ok := pref[q.QuotaKey]; !ok {
-				continue
-			}
-		}
-		used := 100 - *q.Percent
-		if used < 0 {
-			used = 0
-		}
-		if used > 100 {
-			used = 100
-		}
-		return &used
-	}
-	return nil
-}
-
-func primaryWeeklyKeys(provider string) []string {
-	switch strings.ToLower(strings.TrimSpace(provider)) {
-	case "anthropic", "claude":
-		return []string{"seven_day"}
-	case "codex", "kimi":
-		return []string{"code_week"}
-	case "xai", "grok":
-		return []string{"weekly_limit"}
-	default:
-		return nil
-	}
 }
 
 func (s *Service) listAuths(tenantID string) []*coreauth.Auth {
